% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_onco_xref.R
\name{get_gwas_data}
\alias{get_gwas_data}
\title{Function that retrieves gwasOncoX data from Google Drive}
\usage{
get_gwas_data(
  cache_dir = NA,
  force_download = F,
  cancer_only = T,
  build = "grch38",
  data_type = "vcf"
)
}
\arguments{
\item{cache_dir}{Local directory for data download}

\item{force_download}{Logical indicating if local cache should be overwritten
(set to TRUE to force download if file exists in cache)}

\item{cancer_only}{retrieve data relevant for cancer phenotypes only}

\item{build}{genome assembly (grch37/grch38)}

\item{data_type}{type of data to be retrieved (bed, vcf, rds)}
}
\description{
Function that retrieves gwasOncoX data from Google Drive
}
\keyword{internal}
