% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variants.R
\name{get_variants}
\alias{get_variants}
\title{Get GWAS-associated variants}
\usage{
get_variants(cache_dir = NA, overwrite = F, cancer_only = T)
}
\arguments{
\item{cache_dir}{Local directory for data download}

\item{overwrite}{Logical indicating if local cache should be overwritten
(set to TRUE to re-download if file exists in cache)}

\item{cancer_only}{logical indicating retrieval of all variants versus
cancer-associated variants only}
}
\value{
a data frame with varions annotations (phenotype ID's,
association p-value, risk alleles, PMIDs) for GWAS variants
}
\description{
Downloads and returns a dataset with variants associated with
disease phenotypes - as discovered from genome-wide association studies.
}
