% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vcf.R
\name{get_vcf}
\alias{get_vcf}
\title{Get VCF file with GWAS variants}
\usage{
get_vcf(cache_dir = NA, overwrite = F, build = "grch37", cancer_only = T)
}
\arguments{
\item{cache_dir}{Local directory for data download}

\item{overwrite}{Logical indicating if local cache should be overwritten
(set to TRUE to re-download if file exists in cache)}

\item{build}{genome assembly (grch37/grch38)}

\item{cancer_only}{logical indicating retrieval of all variants versus
cancer-associated variants only}
}
\value{
A data frame with file download information and metadata
}
\description{
Downloads a tabix-indexed VCF file (\code{.vcf.gz} and \code{.vcf.gz.tbi}) with variants
associated with disease phenotypes (as found in genome-wide association studies)
The GWAS_HIT element of the INFO column has the following format:
rsid|risk_allele|pmid|tag_snp|p_valule|efo_id
}
